+++ patches/src/main/kotlin/app/morphe/patches/tiktok/interaction/blockauthor/BlockAuthorPatch.kt
@@ -32,7 +32,7 @@ val blockAuthorPatch = bytecodePatch(
 ) {
     dependsOn(settingsPatch, sharedExtensionPatch)
 
-    compatibleWith(*AppCompatibilities.tiktok4623())
+    compatibleWith(*AppCompatibilities.tiktok4703())
 
     execute {
         SettingsStatusLoadFingerprint.method.addInstruction(