+++ patches/src/main/kotlin/app/morphe/patches/tiktok/misc/navigation/FeedTabNavigationPatch.kt
@@ -25,7 +25,7 @@ val feedTabNavigationPatch = bytecodePatch(
 ) {
     dependsOn(settingsPatch, sharedExtensionPatch)
 
-    compatibleWith(*AppCompatibilities.tiktok4623())
+    compatibleWith(*AppCompatibilities.tiktok4703())
 
     execute {
         SettingsStatusLoadFingerprint.method.addInstruction(