+++ patches/src/main/kotlin/app/morphe/patches/tiktok/promobanners/HideFloatingPromotionsPatch.kt
@@ -47,7 +47,7 @@ val hideFloatingPromotionsPatch = bytecodePatch(
     default = true,
 ) {
     dependsOn(settingsPatch, sharedExtensionPatch)
-    compatibleWith(*AppCompatibilities.tiktok4623())
+    compatibleWith(*AppCompatibilities.tiktok4703())
 
     execute {
         SettingsStatusLoadFingerprint.method.addInstruction(